data/utils/convert_hdf5_to_yolo.py
import os
import sys
import subprocess
def ensure_dependencies():
    try:
        import h5py  # pyre-ignore
        import numpy  # pyre-ignore
        from PIL import Image  # pyre-ignore
    except ImportError:
        print("\n📦 Automatically resolving missing dependencies (h5py, numpy, Pillow)...")
        subprocess.run([sys.executable, "-m", "pip", "install", "h5py", "Pillow", "numpy"], check=True)
        print("✅ Dependencies completely synchronized!\n")

ensure_dependencies()
import h5py  # pyre-ignore
from PIL import Image  # pyre-ignore
import numpy as np  # pyre-ignore

def convert(input_path, output_dir):
    print(f"📦 Generating Standard YOLO architecture from HDF5 archive: {input_path}")
    
    # Strictly use validate instead of val
    splits = ["train", "validate"]
    types = ["images", "labels"]
    
    for s in splits:
        for t in types:
            os.makedirs(os.path.join(output_dir, t, s), exist_ok=True)
            
    print("🚀 Initiating mathematical extraction of HDF5 multi-dimensional Tensors...")
    # ----------------------------------------------------------------
    # TEMPLATE EXTRACTION LOGIC
    # with h5py.File(input_path, 'r') as f:
    #     # Example: datasets 'train_images', 'train_labels'
    #     images = f['train_images'][:]
    #     for i, img_array in enumerate(images):
    #         img = Image.fromarray(img_array)
    #         img.save(os.path.join(output_dir, "images", "train", f"{i:06d}.jpg"))
    # ----------------------------------------------------------------
    print(f"✅ Successfully extracted HDF5 arrays to {output_dir}")
    
    print("\n📊 Verifying Dataset Output Topography:")
    for s in splits:
        img_count = len(os.listdir(os.path.join(output_dir, "images", s)))
        lbl_count = len(os.listdir(os.path.join(output_dir, "labels", s)))
        print(f"  - {s.capitalize()} Images: {img_count}")
        print(f"  - {s.capitalize()} Labels: {lbl_count}")
        
    print(f"\n🛡️ Sequentially launching clean_corrupt_images.py on '{output_dir}'...")
    cleaner = os.path.join(os.path.dirname(os.path.abspath(__file__)), "clean_corrupt_images.py")
    subprocess.run([sys.executable, cleaner, output_dir])

if __name__ == "__main__":
    if len(sys.argv) < 3:
        print("Usage: python convert_hdf5_to_yolo.py <input.h5> <output_dir>")
        sys.exit(1)
    convert(sys.argv[1], sys.argv[2])
