requirements.txt
torch==2.4.1+cu121
torchvision==0.19.1+cu121
torchmetrics>=1.0.0
pillow>=9.0.0
numpy>=1.21.0
pyyaml>=6.0
psutil>=5.8.0
pandas>=1.3.0
pyarrow>=10.0.0
scipy>=1.7.0
onnx>=1.12.0
onnxsim>=0.4.0
onnxconverter-common>=1.9.0
tqdm>=4.64.0
torch-directml
tensorflow-cpu>=2.10.0
ultralytics>=8.0.0